Issue #1843: Turn SessionManager into a facade and move functionality to LegacySessionManager.
